Add rudimentary UI support for QV plugin.
Since QV needs to be paired up with some other plugin that protects against sybil attacks, it will typically have a predecessor plugin (e.g. the Civic Gateway plugin)

In milestone 3 of the QV work, there will be a way to configure this via the UI.

In the meantime, we use the heuristic that the existing plugin should be used as the predecessor.

If this is not desired, for any reason, then the workaround is to move back to "default" first.

We also only enable default configuration of the quadratic plugin, resulting in a vote weight that is the square root of the token balance.