Bugfix/delegation fix disable batch (#75)
* Fix delegation with plugins - currently TORs are created on voting, but VWRs are not. This means plugins with non-emphemeral VWRs will not work for delegates who have not themselves onboarded, in "batch voting" mode.

After discussion, this will be solved by filtering those out of the voting batch, but that will be done in a separate PR. For now, batch voting mode is disabled for these plugins. Additionally, a bug is fixed that shows the combined votes on the UI when "Yourself only" is picked

* Fix delegation with plugins - currently TORs are created on voting, but VWRs are not. This means plugins with non-emphemeral VWRs will not work for delegates who have not themselves onboarded, in "batch voting" mode.

After discussion, this will be solved by filtering those out of the voting batch, but that will be done in a separate PR. For now, batch voting mode is disabled for these plugins. Additionally, a bug is fixed that shows the combined votes on the UI when "Yourself only" is picked